refactor(runtime): retire active-full compact
Generated-by: Codex
