+++ packages/runtime-host/src/server/execution-model-composition.ts
@@ -388,7 +388,6 @@ export async function createHostAiSdkBackend(input: HostAiSdkBackendInput): Prom
         recordHistoryCompactCheckpoint: input.context.recordHistoryCompactCheckpoint,
         loadTurnRuntimeEvents: input.context.loadTurnRuntimeEvents,
         allowMidTurnHistoryCompaction: input.context.allowMidTurnHistoryCompaction,
-        recordActiveFullCompactBlock: input.context.recordActiveFullCompactBlock,
         recordSemanticCompactBlock: input.context.recordSemanticCompactBlock,
         recordRunTrace: input.context.recordRunTrace,
         ...(commitRunComposition
