+++ packages/runtime/src/__tests__/ai-sdk-backend.test.ts
@@ -58,7 +58,6 @@ import {
 import { buildDefaultContextBudgetPolicy } from '../context-budget-policy.js';
 import { memoryArtifactStore } from './memory-artifact-store.js';
 import { buildRuntimeEventModelReplayPlan, buildSteeringEnvelope } from '../model-history.js';
-import type { ActiveFullCompactBlock } from '../active-full-compact.js';
 import type { SemanticCompactBlock } from '../semantic-compact.js';
 import { HistoryCompactSummarizerError } from '../history-compact-summarizer.js';
 import type { SandboxDiagnosticsSnapshot } from '../sandbox/diagnostics.js';
@@ -8791,168 +8790,6 @@ describe('AiSdkBackend usage telemetry', () => {
     ) as { contextBudget?: Record<string, unknown> } | undefined;
     assert.equal(usageMessage?.contextBudget?.activeSupersededToolResults, 1);
     assert.equal(usageMessage?.contextBudget?.activeDuplicateToolResults, undefined);
-  });
-
-  test('active full compact sees the fresh tool result before active tool-result prune', async () => {
-    const durable = durableTurnHarness('turn-1', 'hi');
-    const messages: unknown[] = [];
-    const events: SessionEvent[] = [];
-    const recordedBlocks: ActiveFullCompactBlock[] = [];
-    const largeBody = 'ACTIVE_FULL_COMPACT_RAW_TOOL_OUTPUT'.repeat(200);
-    let streamCalls = 0;
-    let secondProviderRequestSawRecordedBlock = false;
-    const model = new MockLanguageModelV4({
-      doStream: async () => {
-        streamCalls += 1;
-        if (streamCalls === 2) {
-          secondProviderRequestSawRecordedBlock = recordedBlocks.length === 1;
-        }
-        const chunks: LanguageModelV4StreamPart[] =
-          streamCalls === 1
-            ? [
-                { type: 'stream-start', warnings: [] },
-                {
-                  type: 'tool-call',
-                  toolCallId: 'tool-1',
-                  toolName: 'Read',
-                  input: JSON.stringify({ path: 'notes.md' }),
-                },
-                {
-                  type: 'finish',
-                  finishReason: { unified: 'tool-calls', raw: 'tool_calls' },
-                  usage: {
-                    inputTokens: { total: 1, noCache: 1, cacheRead: 0, cacheWrite: 0 },
-                    outputTokens: { total: 1, text: 1, reasoning: 0 },
-                  },
-                },
-              ]
-            : [
-                { type: 'stream-start', warnings: [] },
-                {
-                  type: 'finish',
-                  finishReason: { unified: 'stop', raw: 'stop' },
-                  usage: {
-                    inputTokens: { total: 1, noCache: 1, cacheRead: 0, cacheWrite: 0 },
-                    outputTokens: { total: 1, text: 1, reasoning: 0 },
-                  },
-                },
-              ];
-        return {
-          stream: simulateReadableStream({ chunks, initialDelayInMs: null, chunkDelayInMs: null }),
-        };
-      },
-    });
-    const backend = createTestAiSdkBackend({
-      sessionId: 'session-1',
-      header: header(),
-      appendMessage: async (message) => {
-        messages.push(message);
-      },
-      connection: connection(),
-      apiKey: 'sk-test',
-      modelId: 'mock-model-id',
-      modelFactory: () => model,
-      tools: [
-        {
-          name: 'Read',
-          description: 'Read description',
-          parameters: z.object({ path: z.string() }),
-          impl: async () => ({ body: largeBody }),
-        },
-      ],
-      contextBudget: {
-        charsPerToken: 1,
-        activeToolResultPrune: { enabled: true, maxCurrentResultEstimatedTokens: 1 },
-        activeFullCompact: {
-          enabled: true,
-          minStepNumber: 1,
-          minRecentMessages: 0,
-          maxActiveEstimatedTokens: 1,
-          highWaterRatio: 0.1,
-          maxSummaryEstimatedTokens: 512,
-        },
-      },
-      toolResultArchive: testToolResultArchive({
-        archiveToolResult: async () => ({ artifactId: 'artifact-tool-1' }),
-      }),
-      loadTurnRuntimeEvents: durable.loadTurnRuntimeEvents,
-      newId: idGenerator(),
-      now: monotonicClock(),
-      recordActiveFullCompactBlock: (block) => {
-        recordedBlocks.push(block);
-      },
-    });
-
-    for await (const event of backend.send(durable.input())) {
-      durable.record(event);
-      events.push(event);
-    }
-    await Promise.resolve();
-
-    assert.equal(streamCalls, 2);
-    assert.equal(recordedBlocks.length, 1);
-    assert.equal(secondProviderRequestSawRecordedBlock, true);
-    assert.equal(recordedBlocks[0]?.kind, 'maka.active_full_compact_block');
-    assert.equal(recordedBlocks[0]?.turnId, 'turn-1');
-    assert.equal((recordedBlocks[0]?.sourceRefs.length ?? 0) > 0, true);
-    assert.doesNotMatch(JSON.stringify(recordedBlocks[0]), /artifact-tool-1/);
-    const secondPromptMessages = model.doStreamCalls[1]?.prompt ?? [];
-    const secondPrompt = JSON.stringify(
-      model.doStreamCalls[1]?.prompt.map((message) => ({
-        role: message.role,
-        content: message.content,
-      })),
-    );
-    assert.match(secondPrompt, /maka_active_full_compact_block/);
-    assert.equal(
-      secondPromptMessages.some(
-        (message) =>
-          message.role === 'user' &&
-          JSON.stringify(message.content).includes('maka_active_full_compact_block'),
-      ),
-      true,
-    );
-    assert.equal(
-      secondPromptMessages.some(
-        (message) =>
-          message.role === 'system' &&
-          JSON.stringify(message.content).includes('maka_active_full_compact_block'),
-      ),
-      false,
-    );
-    assert.doesNotMatch(secondPrompt, /artifact-tool-1/);
-    assert.equal(secondPrompt.includes('ACTIVE_FULL_COMPACT_RAW_TOOL_OUTPUT'), false);
-    assert.doesNotMatch(secondPrompt, /providerSourceIds=/);
-    assert.doesNotMatch(secondPrompt, /bodySha256=/);
-    assert.doesNotMatch(secondPrompt, /source\(kind=/);
-
-    const usageMessage = messages.find(
-      (message) => (message as { type?: string }).type === 'token_usage',
-    ) as { contextBudget?: Record<string, unknown> } | undefined;
-    const usageEvent = events.find((event) => event.type === 'token_usage') as
-      | (Extract<SessionEvent, { type: 'token_usage' }> & {
-          contextBudget?: Record<string, unknown>;
-        })
-      | undefined;
-    for (const contextBudget of [usageMessage?.contextBudget, usageEvent?.contextBudget]) {
-      assert.equal(contextBudget?.activePrunedToolResults, undefined);
-      const decisions = contextBudget?.compactionDecisions as
-        | Array<Record<string, unknown>>
-        | undefined;
-      assert.equal(
-        decisions?.some(
-          (decision) =>
-            decision.boundaryKind === 'activeFullCompact' && decision.decision === 'replaced',
-        ),
-        true,
-      );
-      assert.equal(typeof contextBudget?.highWaterRequestShapeHashBefore, 'string');
-      assert.equal(typeof contextBudget?.highWaterRequestShapeHashAfter, 'string');
-      assert.notEqual(
-        contextBudget?.highWaterRequestShapeHashAfter,
-        contextBudget?.highWaterRequestShapeHashBefore,
-      );
-    }
   });
 
   test('normalizes cache and reasoning tokens to messages, events, and telemetry', async () => {
