+++ packages/runtime/src/__tests__/semantic-compact.test.ts
@@ -1,7 +1,7 @@
 import assert from 'node:assert/strict';
 import { describe, test } from 'node:test';
 import type { ModelMessage } from '../model-protocol.js';
-import { buildActiveCompactionHeadAnchor } from '../active-full-compact.js';
+import { buildActiveCompactionHeadAnchor } from '../active-compaction-kernel.js';
 
 import {
   renderSemanticCompactBlock,
