+++ packages/runtime/src/request-projection.ts
@@ -27,9 +27,9 @@ export type RequestProjectionStage = (
  * Order is a contract: mid-turn capacity compaction runs first among the
  * message-shaping hooks so every later mechanism operates on (and re-converges
  * onto) its projection — active tool-result pruning re-archives large tool
- * results in the rebuilt tail, and semantic/active-full compaction sees the
+ * results in the rebuilt tail, and semantic compaction sees the
  * already-compacted messages. On a step where the capacity hook replaced the
- * request, semantic/active-full compaction yields (see AiSdkBackend.send) so
+ * request, semantic compaction yields (see AiSdkBackend.send) so
  * two summarizers never run for one step.
  *
  * Every hook here only SHAPES the projection. The pass/terminate capacity
@@ -42,13 +42,13 @@ export function composeRequestProjection(
   toolAvailability: RequestProjectionStage | undefined,
   midTurnCapacityCompact: RequestProjectionStage | undefined,
   activeToolResultPrune: RequestProjectionStage | undefined,
-  activeFullCompact?: RequestProjectionStage | undefined,
+  activeCompaction?: RequestProjectionStage | undefined,
 ): RequestProjectionStage | undefined {
   const hooks = [
     toolAvailability,
     midTurnCapacityCompact,
     activeToolResultPrune,
-    activeFullCompact,
+    activeCompaction,
   ].filter(Boolean) as RequestProjectionStage[];
   if (hooks.length === 0) return undefined;
   return async (context: RequestProjectionContext): Promise<RequestProjection | undefined> => {
